metrics/requirements.txt
prometheus_client==0.17.1
requests==2.31.0
python-telegram-bot==20.7 